More tests and test refactoring
